Version 1.0
Version1.0. Makes a html file from a directory of .url and .website
files. By default it looks in IE's shortcut directory and puts the
resulting html file on the desktop, both of these can be overridden with
the command line. The html file will have a link for each shortcut in
the specified directory, the link will be named the same as the original
shortcut and will go to the same place as the shortcut. By default it
will also parse through subdirectories but this can be turned off with
the command line.